Support reboot via WEB api

+++ reboot.py
@@ -10,9 +10,20 @@ import gateway
 from gateway import die
 
 
-gw = gateway.Gateway()
+gw = gateway.Gateway(detect_ssh = False)
 
-print("Send command...")
-gw.run_cmd("reboot")
+ssh = gw.detect_ssh(verbose = 1, interactive = True)
+if ssh > 0:
+  print('Send command "reboot" via SSH ...')
+  gw.run_cmd("reboot")
+else:
+  if not gw.stok:
+    gw.web_login()
+  print('Send command "reboot" via WEB API ...')
+  if not gw.reboot_device():
+    die('Can\'t run reboot command.')
+
+if not gw.wait_shutdown(10):
+  die('The "reboot" command did not shutdown the device.')
 
 print("Reboot activated!")
